Move plugins.
