+++ src/plugins/mod.rs
@@ -1,0 +1,4 @@
+//! comrak plugins we need
+
+mod slides;
+pub use slides::SlideAdapter;
